Add support for denied edges in the denylist
This change adds support for denying of hotspot-hotspot edges when the
denylist system considers an edge invalid because of one of its
classifiers but is unable to assign blame to a specific hotspot. This
allows the denylist to deny earnings for bad edges without accidentally
denying hotspots that are simply adjacent or enmeshed in a cluster of
badly behaving hotspots, or anoymous radio repeater devices.

Hotspot edges are not directional, so A->B is considered the same as
B->A. To accomplish this, the hotspot keys are always lexiographically
ordered when constructing the hash to check the denylist.