Mj/footfall boosting (#761)
* add footfall data to boosting flow

footfall and urbanization are considered together for determining the
multiplier applied to a hex.

* require passing a valid hex cell when asking about a cell

* Fix hex_assignment

* Fix tests

* Fix formatting

* Fix footfall Assignment to A for coverage tests

* Provide mocked hex assignments struct

Test that don't care about the specific values or underlying structures
of HexAssignments can use this mock. The only option available is to
return the best possible multiplier for a hex, hopefully behaving in a
way that removes hex boosting from the equation.

* restructure HexBoosting

The top level HexBoosting struct is the wrapper for asking internal
DiskTree about Assignments. For debugability, the information inside a
DiskTree is turned into an enum mirroring the verbage in HIP-103, then
get's turned into an Assignment.

* use helper to make hex boost data with settings path

* keep naming consistent between structs

* no longer only urbanization hexes, all unassigned hexes

* Update mobile_verifier/src/boosting_oracles/mod.rs

Co-authored-by: Matthew Plant <matty@nova-labs.com>

* Implement HexAssignment for both of the data sets independently

* add file-store dump for coverage object

* add error context for keypair deserializing

* add dump mobile rewards command

groups mobile rewards from a file into a valid json object and prints

* Pull calculations out of code into structs

Co-authored-by: macpie <macpie@users.noreply.github.com>

* more nonzero use

* format

* update helium-proto

* Change proto back to master

---------

Co-authored-by: Michael Jeffrey <michaeldjeffrey@gmail.com>
Co-authored-by: Matthew Plant <matty@nova-labs.com>
Co-authored-by: Brian Balser <bbalser@nova-labs.com>