Mobile Verifier Burner Txn Error check (#906)
* Make SolanaNetwork implement Clone so we can get a copy for a spawned task

* Add retry settings to mobile verifier burner

* Fallback to confirming transactions on all errors

If a submitted transaction does not come back as a defacto success, we spawn a process that checks a txn made it to the chain by signature. By default, we will retry 5 times a minute apart, before considering the txn an actual failure to be tried again later.

If during the checking we find the txn on chain, we do the same logic as if it had passed the first time.

* consistent txn in logs

* handle success case first

* prefix with failed so we know what we're retrying

* block on txn confirmation

mobile packet verifier is not handling many burn txns, we can block on confirming before attempting to process the next txn.

Txn tracking by signature and block-height will come in a later update.