Hotspot Banning (#984)
* Mobile Banning

Process ban reports in the mobile verifier.
Next up, making sure rewards are not paid out for banned radios.

* Assume the BannedRadio into banning

SP bans will be going away someday :fingers-crossed:. 

Right now, the difference is SP bans set an oracle boost multiplier to `0`, resulting in no rewards, but they are still written to the rewards report.

New banning completely ignores processing the radio.

* Add non-blocking message consumption for testing
* rewrite tests to not block in file sink writing
* Add verified ban source
* mobile-packet-verifier ban code
* Add mpv banning to settings
* Add banning managed task to daemon
* add banned_radios to accumulate
* Add comments about why we track POC bans
* make mobile-verifier banning look more like mobile-packet-verifier
* Cleanup bans at the end of rewarding
* add status metric when writing verified report
* Update proto fields 

- `notes` -> `message`
- Add more specific BanReason

* add forgotten biased selector for mpv

if there are multiple files available and we're trying to shutdown, we always want to shut down non-randomly.

* tokio::select! that check for shutdown should be biased
* make ban_type check more explicit

If more ban_types are added in the future we should update the queries in the places that care about the specific types of bans, as we ingest all of them.

* Make sure mobile packet verifier does not ingest invalid ban requests
* Make sure a ban was received within the time of rewarding or file processing

- for rewarding, we want to make sure they were banned before or during the reward epoch. This also makes sure we can rerun previous rewards
- for mobile-packet-verifier, we want to make sure a hotspot is banned for the time that the currently processing file came in

* If the FileInfoPoller is dropped we want to err the ingestor process for restarting

change log messages to match other ingester

* Move sp_boosted_bans into banning namespace

- Keep 7 days of history for bans like sp_boost_bans
- consolidate clear_bans for rewarder

* Add 7 day history for bans to mobile-packet-verifier
* Be more idempotent with radio bans

Store all incoming bans with keyed by their (hotspot_pubkey, received_timestamp).

When we query for banned radios, only look for the latest record for a given hotspot_pubkey at the given time, then filter if it's a ban_type we care about. (These are opposite in mobile-verifier and mobile-packet-verifier).

Intead of removing bans, we mark them as expiring at the time the unban was received.

This means we can process files from previous days, and if a currently unbanned radio would have been banned, it will still be considered so.

* More explicitly spell out how we handle bans on the edge cases
* Fix past-future ban tests

The query for dealing with bans cares more about the epoch rather than a moment in time. It's query is written thusly, and we need to adjust how we insert our test values to match.

* Go back to proto master

ban_key -> ban_pubkey