Remove Generic Associated Error Types (#992)
In every occurence where we have a GAT error type, the ability to change the Error is not used. As all of the touched traits are for API Clients, and they all return ClienError, we can reduce the amount of generic constraints we need to type by removing them.

In a case where we do want to return an error, it is a simple process to return an applicable ClientError variant.