Update AWS Dependencies (Welcome to the Jungle) (#1019)
* Working branch with aws_sdk_s3 latest

* Box s3 error 

It might be larger than the other error variants enough that clippy is complaining about it.

* Hard code aws deps to same as file_store

I think it would be a good idea to roll this into file-store at some point to remove the need to keep them in line.

* Force path style when using s3 locally with minio/localstack

Path Style requests are being slowly deprecated. Fortunately, the s3_config still provides a way to override this setting.

This is mostly for testing locally with minio or localstack, as you would need to setup some DNS rules for your localhost to get virtual-hosted style paths to work.

https://docs.aws.amazon.com/AmazonS3/latest/userguide/VirtualHosting.html

* Update db_store rds_iam_token generation

This commit exists with both version so they can be compared in history, and in to show that there was a time where it was known that both functions generated the same token.

* Cleanup iam_auth_pool

- remove old aws crates
- consolidate signing errors
- remove invalid auth token error, we construct the url almost statically

* remove unused aws-types crate

* Format cargo.toml

* Update rust toolchain for required aws sdk msrv

and update clippy warnings

* Working deps lock file

* Move AWS deps to root

Group them all together to try and reduce the amount of version hunting we need to do next time we need to update a dep.

* Update helium-lib

* Update to rust 1.87

* Update to rust 1.88

* Add clippy config to ignore tonic::Status lint errors

We can remove this lint once # 
https://github.com/hyperium/tonic/issues/2253
 has a PR branch and is merged. And we can update from tonic `0.10` to 
 `0.13.x`.

* Back to mainline crypto-rs and helium-lib

* remove load timeout settings from filestore

It has been removed in updated version of aws sdks, there are ways to
handle it, but they're much more involved.

We can test with updated versions of loading credentials, and it we're
still getting timeout failures for price tracker, we can consider adding
it back in with the new style.

* Use recommendation from deprecation message

We have an array, and try_from works on a slice of u8.

* Update helium-proto tonic imports

tonic is no longer exported from helium-proto, we define it as a dep,
and it picks up it's version from helium-lib.

* update http version

it will pick up it's version from the re-export of http-serde

* remove use of deprecated mobile config info calls

* use info_v2 as info has been deprecated for a while

* use enum getter from prost

* format cargo toml

* clippy updates, format strings and clones

* update itertools

* working Cargo.lock file

* update to cargo 1.89.0

- redundant return statements
- confusing elided lifetime
- unused test struct

* use cmake compiler when building aws-lc-sys

This has a version of cc that is not bugged.

* need cmake in the docker container to build with it

* use cmake when building debian packages as well

* set AWS_LC_SYS_CMAKE_BUILDER as an env variable for everything

* install cmake at the right stage

* mark Aws Error in FileStore::Error as #[source]

---------

Co-authored-by: Brian Balser <bbalser@nova-labs.com>