Crypto Provider fix 🤞  (#1051)
* update backon

If we let it update itself between 0.4 and 0.5 there is a breaking
change where a sleeper is no longer provided.

Forcing to ^1 fixes the problem.

* use sqlx version from workspace

* add tls_init crate

This crate attempts to lock down which version of rustls crypto provider
we use to be only aws-lc-rs.

The ctor crate allows us to define a constructor that runs when a
package is loaded. And it ensures the constructor is only run a single
time no matter how many times it's included.

This is the alternative to littering all our main functions with
`rustls::crypto::aws_lc_rs::default_provider().install_provider()`
calls.

Because the linker will attempt to remove unused code, and there are no
direct calls to this function, we use `extern crate tls_init;` wherever
we want to ensure we have the `aws-lc-rs` backend. Basically, it tells
the compiler, this dependency is definitely used even though you might
not see it.

* add tls-init wherever we think there might be tls needed

* test macro for downstream crates

include this macro in crates that need to ensure tls is provided by
aws-lc-rs. Hopefully this will catch potential conflicts if a dependency
is included that enables something other than aws-lc.

* add docs about crate and macro